+++ app/src/main/java/de/baumann/weather/helper/CustomListAdapter.java
@@ -16,15 +16,17 @@ public class CustomListAdapter extends ArrayAdapter<String> {
     private final Activity context;
     private final String[] itemname;
     private final String[] itemURL;
+    private final String[] itemDES;
     private final Integer[] imgid;
 
-    public CustomListAdapter(Activity context, String[] itemname, String[] itemURL,Integer[] imgid) {
+    public CustomListAdapter(Activity context, String[] itemname, String[] itemURL, String[] itemDES, Integer[] imgid) {
         super(context, R.layout.list_item, itemname);
         // TODO Auto-generated constructor stub
 
         this.context=context;
         this.itemname=itemname;
         this.itemURL=itemURL;
+        this.itemDES=itemDES;
         this.imgid=imgid;
     }
 
@@ -35,10 +37,12 @@ public class CustomListAdapter extends ArrayAdapter<String> {
         TextView txtTitle = (TextView) rowView.findViewById(R.id.item);
         ImageView imageView = (ImageView) rowView.findViewById(R.id.icon);
         TextView extratxt = (TextView) rowView.findViewById(R.id.textView1);
+        TextView destxt = (TextView) rowView.findViewById(R.id.textView0);
 
         txtTitle.setText(itemname[position]);
         imageView.setImageResource(imgid[position]);
         extratxt.setText(itemURL[position]);
+        destxt.setText(itemDES[position]);
         return rowView;
 
     }